PersonalHotGames/requirements.txt
python-dateutil
requests
requests_cache
xmltodict